[ICC-315] ♻️ refactor: util → document 모듈 리네이밍 및 deprecated UpdateLog 삭제
- util 모듈을 document로 리네이밍 (패키지: com.icc.qasker.document)
- deprecated UpdateLog 관련 코드 7개 파일 삭제
- HelloController를 app 모듈로 이동
- document-impl에서 불필요한 JPA/MySQL/Web/Swagger 의존성 제거
