--- modules/util/api/src/main/java/com/icc/qasker/util/dto/request/UpdateLogRequest.java
@@ -1,3 +1,0 @@
-package com.icc.qasker.util.dto.request;
-
-public record UpdateLogRequest(String updateText) {}
