--- modules/util/api/src/main/java/com/icc/qasker/util/dto/response/UpdateLogResponse.java
@@ -1,9 +1,0 @@
-package com.icc.qasker.util.dto.response;
-
-import java.time.Instant;
-import java.util.List;
-
-public record UpdateLogResponse(List<UpdateLog> updateLogs) {
-
-  public record UpdateLog(Instant dateTime, String updateText) {}
-}
