--- modules/util/impl/build.gradle
@@ -1,64 +1,0 @@
-plugins {
-    // java-library: api/implementation 의존성 구분을 지원하는 라이브러리 모듈용 플러그인
-    // 예: api로 선언한 의존성은 이 모듈을 사용하는 소비자에게 전이됨,
-    //     implementation으로 선언하면 내부에서만 사용되고 소비자에게 노출되지 않음
-    id 'java-library'
-}
-
-dependencies {
-
-    // ────────────────────────────────
-    // 프로젝트 모듈
-    // ────────────────────────────────
-    // global: 공통 예외 처리, 공통 응답 DTO 등 프로젝트 전역 유틸리티 모듈
-    // 예: GlobalException, ApiResponse<T> 등 전역 클래스를 util-impl에서 참조
-    implementation project(":global")
-    // util-api: 유틸리티 관련 인터페이스·DTO 정의 모듈
-    // 예: UtilController 인터페이스, 응답 DTO를 구현체에서 import
-    implementation project(":util-api")
-
-    // ────────────────────────────────
-    // JODConverter: LibreOffice 기반 문서 변환
-    // ────────────────────────────────
-    // 예: DocumentConverter.convert(inputFile).to(outputFile).execute()로 PPT/DOCX → PDF 변환
-    //     Spring Boot Starter가 OfficeManager, DocumentConverter 빈을 자동 등록
-    implementation "org.jodconverter:jodconverter-local-lo:4.4.9"
-    implementation "org.jodconverter:jodconverter-spring-boot-starter:4.4.9"
-
-    // ────────────────────────────────
-    // Micrometer: Prometheus 커스텀 메트릭 (문서 변환 Timer)
-    // ────────────────────────────────
-    implementation "io.micrometer:micrometer-core"
-
-    // ────────────────────────────────
-    // 스프링부트 Web & JPA
-    // ────────────────────────────────
-    // Spring MVC 기반 웹 애플리케이션 구성
-    // 예: @RestController로 유틸리티 API 엔드포인트 구현
-    implementation "org.springframework.boot:spring-boot-starter-web"
-    // Spring Data JPA: JPA 기반 리포지토리 자동 생성, 엔티티 매핑 지원
-    // 예: JpaRepository 인터페이스 선언만으로 CRUD 메서드가 자동 구현됨
-    implementation "org.springframework.boot:spring-boot-starter-data-jpa"
-    // MySQL JDBC 드라이버: 런타임에 DB 연결 시 사용 (컴파일 시에는 불필요)
-    // 예: JPA가 내부적으로 이 드라이버를 통해 MySQL 서버와 TCP 커넥션을 맺음
-    runtimeOnly "com.mysql:mysql-connector-j"
-    
-
-    // ────────────────────────────────
-    // API 문서 (Swagger)
-    // ────────────────────────────────
-    // SpringDoc OpenAPI: API 문서화 어노테이션(@Operation, @Schema) + Swagger UI 제공
-    // 예: 서버 기동 후 /swagger-ui/index.html 접속하면 API 목록·테스트 UI가 제공됨
-    implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.8"
-
-    // ────────────────────────────────
-    // 테스트
-    // ────────────────────────────────
-    testImplementation "org.springframework.boot:spring-boot-starter-test"
-    testImplementation platform("org.junit:junit-bom:5.12.2")
-    testRuntimeOnly "org.junit.platform:junit-platform-launcher"
-}
-
-test {
-    useJUnitPlatform()
-}
