--- modules/util/impl/src/main/java/com/icc/qasker/util/entity/UpdateLog.java
@@ -1,27 +1,0 @@
-package com.icc.qasker.util.entity;
-
-import com.icc.qasker.global.entity.CreatedAt;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Builder
-public class UpdateLog extends CreatedAt {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(length = 255, nullable = false)
-  private String updateText;
-}
